+++ src/haven/purus/Config.java
@@ -98,6 +98,8 @@ public class Config {
 	public static Setting<Boolean> toggleDlight = new Setting<>("toggleDlight", false);
 	public static Setting<Boolean> flavorObjsVisual = new Setting<>("flavorObjsVisual", false);
 	public static Setting<Boolean> flavorObjsAll = new Setting<>("flavorObjsAudial", false);
+	public static Setting<Boolean> animalRadiuses = new Setting<>("animalRadiuses", false);
+
 
 	public static Setting<Boolean> hideTrees = new Setting<>("hideTrees", true);
 	public static Setting<Boolean> hideBushes = new Setting<>("hideBushes", true);
@@ -132,6 +134,7 @@ public class Config {
 	public static KeyBinding kb_hidetoggle = KeyBinding.get("kb_hidetoggle", KeyMatch.forcode(KeyEvent.VK_H, KeyMatch.C));
 	public static KeyBinding kb_camswitch = KeyBinding.get("kb_camswitch", KeyMatch.forcode(KeyEvent.VK_C, KeyMatch.S));
 	public static KeyBinding kb_dlighttoggle = KeyBinding.get("kb_dlighttoggle", KeyMatch.forcode(KeyEvent.VK_N, KeyMatch.C));
+	public static KeyBinding kb_animalradius = KeyBinding.get("kb_animalradius", KeyMatch.forcode(KeyEvent.VK_D, KeyMatch.S));
 
 	public static Setting<Integer> camCycle = new Setting<>("camCycle", 0);
 	public static String[][] camCycles = new String[][]{{"cam", "bad"}, {"cam", "worse"}, {"cam", "follow"}, {"cam", "ortho"}, {"cam", "ortho", "-f"}};